% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{acceptednamesfromtsn}
\alias{acceptednamesfromtsn}
\title{Get accepted names from tsn}
\usage{
acceptednamesfromtsn(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get accepted names from tsn
}
\examples{
\dontrun{
# TSN accepted - good name
acceptednamesfromtsn(tsn='208527')

# TSN not accepted - input TSN is old name
acceptednamesfromtsn('504239')

# raw json
acceptednamesfromtsn(tsn='208527', raw = TRUE)
# raw xml
acceptednamesfromtsn(tsn='208527', wt = "xml", raw = TRUE)
}
}

