% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{anymatchcount}
\alias{anymatchcount}
\title{Get any match count.}
\usage{
anymatchcount(x, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
An integer containing the number of matches the search will return.
}
\description{
Get any match count.
}
\examples{
\dontrun{
anymatchcount(202385)
anymatchcount("dolphin")
}
}

