% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_names.R
\name{common_names}
\alias{common_names}
\title{Get common names from tsn}
\usage{
common_names(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get common names from tsn
}
\examples{
\dontrun{
common_names(tsn=183833)
common_names(tsn=183833, wt = "xml")
}
}

