% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{coremetadata}
\alias{coremetadata}
\title{Get core metadata from tsn}
\usage{
coremetadata(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get core metadata from tsn
}
\examples{
\dontrun{
# coverage and currrency data
coremetadata(tsn=28727)
coremetadata(tsn=28727, wt = "xml")
# no coverage or currrency data
coremetadata(183671)
coremetadata(183671, wt = "xml")
}
}

