% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{coverage}
\alias{coverage}
\title{Get coverge from tsn}
\usage{
coverage(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get coverge from tsn
}
\examples{
\dontrun{
# coverage data
coverage(tsn=28727)
# no coverage data
coverage(526852)
coverage(526852, wt = "xml")
}
}

