% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credibility.R
\name{credibility_ratings}
\alias{credibility_ratings}
\title{Get possible credibility ratings}
\usage{
credibility_ratings(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible credibility ratings
}
\examples{
\dontrun{
credibility_ratings()
credibility_ratings(wt = "xml")
credibility_ratings(raw = TRUE)
credibility_ratings(wt = "xml", raw = TRUE)
}
}

