% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{credibilityrating}
\alias{credibilityrating}
\title{Get credibility rating from tsn}
\usage{
credibilityrating(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get credibility rating from tsn
}
\examples{
\dontrun{
credibilityrating(526852)
credibilityrating(526852, wt = "xml")
credibilityrating(526852, wt = "xml", raw = TRUE)
credibilityrating(526852, raw = TRUE)
}
}

