% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{credibilityratings}
\alias{credibilityratings}
\title{Get possible credibility ratings}
\usage{
credibilityratings(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible credibility ratings
}
\examples{
\dontrun{
credibilityratings()
credibilityratings(wt = "xml")
credibilityratings(raw = TRUE)
credibilityratings(wt = "xml", raw = TRUE)
}
}

