% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{experts}
\alias{experts}
\title{Get expert information for the TSN.}
\usage{
experts(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get expert information for the TSN.
}
\examples{
\dontrun{
experts(tsn = 180544)
experts(180544, wt = "xml")
experts(180544, raw = TRUE)
}
}

