% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_hierarchy.R
\name{full_hierarchy}
\alias{full_hierarchy}
\title{Get full hierarchy from tsn}
\usage{
full_hierarchy(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get full hierarchy from tsn
}
\examples{
\dontrun{
full_hierarchy(tsn=37906)
full_hierarchy(tsn=37906, raw = TRUE)
full_hierarchy(100800, wt = "xml")
}
}

