% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{geographicdivisions}
\alias{geographicdivisions}
\title{Get geographic divisions from tsn}
\usage{
geographicdivisions(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get geographic divisions from tsn
}
\examples{
\dontrun{
geographicdivisions(tsn=180543, config=timeout(3))
}
}

