% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{hierarchy_down}
\alias{hierarchy_down}
\title{Get hierarchy down from tsn}
\usage{
hierarchy_down(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get hierarchy down from tsn
}
\examples{
\dontrun{
hierarchy_down(tsn=161030, config=timeout(3))
}
}

