% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_facet.R
\name{itis_facet}
\alias{itis_facet}
\title{ITIS Solr facet}
\usage{
itis_facet(...)
}
\arguments{
\item{...}{Args passed to \code{\link[solrium]{solr_facet}}}
}
\description{
ITIS Solr facet
}
\examples{
\dontrun{
itis_facet(q = "rank:Species", rows = 0, facet.field = "kingdom")$facet_fields

x <- itis_facet(q = "hierarchySoFar:*$Aves$* AND rank:Species AND usage:valid",
   facet.pivot = "nameWInd,vernacular", facet.limit = -1, facet.mincount = 1,
   rows = 0)
head(x$facet_pivot$`nameWInd,vernacular`)
}
}
