% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_group.R
\name{itis_group}
\alias{itis_group}
\title{ITIS Solr group search}
\usage{
itis_group(...)
}
\arguments{
\item{...}{Args passed to \code{\link[solrium]{solr_group}}}
}
\description{
ITIS Solr group search
}
\examples{
\dontrun{
x <- itis_group(q = "nameWOInd:/[A-Za-z0-9]*[\%20]{1,1}[A-Za-z0-9]*/",
   group.field = 'rank', group.limit = 3)
head(x)
}
}
