% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_highlight.R
\name{itis_highlight}
\alias{itis_highlight}
\title{ITIS Solr highlight}
\usage{
itis_highlight(..., proxy = NULL, callopts = list())
}
\arguments{
\item{...}{Arguments passed on to the \code{params} parameter of
the \code{\link[solrium:solr_highlight]{solrium::solr_highlight()}} function. See \link{solr_fields} for possible
parameters, and examples below}

\item{proxy}{List of arguments for a proxy connection,
including one or more of: url, port, username, password,
and auth. See \code{\link[crul:proxy]{crul::proxy()}} for  help, which is used to
construct the proxy connection.}

\item{callopts}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
ITIS Solr highlight
}
\examples{
\dontrun{
itis_highlight(q = "rank:Species", hl.fl = 'rank', rows=10)
}
}
