% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{itistermsfromcommonname}
\alias{itistermsfromcommonname}
\title{Get itis terms from common names}
\usage{
itistermsfromcommonname(x, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get itis terms from common names
}
\examples{
\dontrun{
itistermsfromcommonname("buya")
itistermsfromcommonname("pum")
}
}

