% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{itistermsfromscientificname}
\alias{itistermsfromscientificname}
\title{Get itis terms from scientific names}
\usage{
itistermsfromscientificname(x, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get itis terms from scientific names
}
\examples{
\dontrun{
itistermsfromscientificname(x="ursidae")
itistermsfromscientificname("Ursus")
}
}

