% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{jurisdiction_values}
\alias{jurisdiction_values}
\title{Get possible jurisdiction values}
\usage{
jurisdiction_values(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get possible jurisdiction values
}
\examples{
\dontrun{
jurisdiction_values(config=timeout(3))
}
}

