% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jurisdiction.R
\name{jurisdictional_origin}
\alias{jurisdictional_origin}
\title{Get jurisdictional origin from tsn}
\usage{
jurisdictional_origin(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get jurisdictional origin from tsn
}
\examples{
\dontrun{
jurisdictional_origin(tsn=180543)
}
}

