% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kingdoms.R
\name{kingdom_name}
\alias{kingdom_name}
\title{Get kingdom names from tsn}
\usage{
kingdom_name(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get kingdom names from tsn
}
\examples{
\dontrun{
kingdom_name(202385, config=timeout(3))
}
}

