% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{kingdom_names}
\alias{kingdom_names}
\title{Get all possible kingdom names}
\usage{
kingdom_names(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Get all possible kingdom names
}
\examples{
\dontrun{
kingdom_names(config=timeout(3))
}
}

