% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_sources.R
\name{other_sources}
\alias{other_sources}
\title{Returns a list of the other sources used for the TSN.}
\usage{
other_sources(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: `FALSE`}
}
\description{
Returns a list of the other sources used for the TSN.
}
\examples{
\dontrun{
other_sources(tsn=182662)
other_sources(tsn=182662, wt = "xml")
}
}
