% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_name.R
\name{rank_name}
\alias{rank_name}
\title{Returns the kingdom and rank information for the TSN.}
\usage{
rank_name(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a data.frame, with rank name and other info
}
\description{
Returns the kingdom and rank information for the TSN.
}
\examples{
\dontrun{
rank_name(tsn = 202385)
}
}
