% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{rank_names}
\alias{rank_names}
\title{Provides a list of all the unique rank names contained in the database and
 their kingdom and rank ID values.}
\usage{
rank_names(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides a list of all the unique rank names contained in the database and
 their kingdom and rank ID values.
}
\examples{
\dontrun{
rank_names(config=timeout(3))
}
}

