% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{record}
\alias{record}
\title{Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.}
\usage{
record(lsid, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{lsid}{lsid for a taxonomic group (character). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Gets the partial ITIS record for the TSN in the LSID, found by comparing the
 TSN in the search key to the TSN field. Returns an empty result set if
 there is no match or the TSN is invalid.
}
\examples{
\dontrun{
record("urn:lsid:itis.gov:itis_tsn:180543", config=timeout(5))
}
}

