% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific_name.R
\name{scientific_name}
\alias{scientific_name}
\title{Returns the scientific name for the TSN. Also returns the component parts
(names and indicators) of the scientific name.}
\usage{
scientific_name(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
a data.frame
}
\description{
Returns the scientific name for the TSN. Also returns the component parts
(names and indicators) of the scientific name.
}
\examples{
\dontrun{
scientific_name(tsn = 531894)
}
}
