% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ritis-package.R
\docType{data}
\name{solr_fields}
\alias{solr_fields}
\title{List of fields that can be used in \link{solr} functions}
\format{A list of length 36}
\source{
\url{https://www.itis.gov/solr_documentation.html}
}
\description{
Each element in the list has a list of length tree, with:
}
\details{
\itemize{
\item field: the field name, this is the name you can use in your
queries
\item definition: the definition of the field
\item example: an example value
}
}
\keyword{data}
