% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{taxon_authorship}
\alias{taxon_authorship}
\title{Returns the author information for the TSN.}
\usage{
taxon_authorship(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Returns the author information for the TSN.
}
\examples{
\dontrun{
taxon_authorship(183671, config=timeout(3))
}
}

