% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsn_by_vernacular_language.R
\name{tsn_by_vernacular_language}
\alias{tsn_by_vernacular_language}
\title{Get tsn by vernacular language}
\usage{
tsn_by_vernacular_language(language, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{language}{A string containing the language. This is a language string,
not the international language code (character)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[crul]{HttpClient}}}
}
\value{
a data.frame
}
\description{
Get tsn by vernacular language
}
\examples{
\dontrun{
tsn_by_vernacular_language(language = "french")
}
}
