% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unacceptability_reason.R
\name{unacceptability_reason}
\alias{unacceptability_reason}
\title{Returns the unacceptability reason, if any, for the TSN.}
\usage{
unacceptability_reason(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Returns the unacceptability reason, if any, for the TSN.
}
\examples{
\dontrun{
unacceptability_reason(tsn = 183671)
}
}

