% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vernacular_languages.R
\name{vernacular_languages}
\alias{vernacular_languages}
\title{Provides a list of the unique languages used in the vernacular table.}
\usage{
vernacular_languages(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a character vector of verncular names
}
\description{
Provides a list of the unique languages used in the vernacular table.
}
\examples{
\dontrun{
vernacular_languages()
}
}
