% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_group.R
\name{itis_group}
\alias{itis_group}
\title{ITIS Solr group search}
\usage{
itis_group(..., proxy = NULL, callopts = list())
}
\arguments{
\item{...}{Arguments passed on to the \code{params} parameter of
the \code{\link[solrium:solr_group]{solrium::solr_group()}} function}

\item{proxy}{List of arguments for a proxy connection,
including one or more of: url, port, username, password,
and auth. See \code{\link[crul:proxy]{crul::proxy()}} for  help, which is used to
construct the proxy connection.}

\item{callopts}{Curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
ITIS Solr group search
}
\examples{
\dontrun{
x <- itis_group(q = "nameWOInd:/[A-Za-z0-9]*[\%20]{1,1}[A-Za-z0-9]*/",
   group.field = 'rank', group.limit = 3)
head(x)
}
}
