% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis_highlight.R
\name{itis_highlight}
\alias{itis_highlight}
\title{ITIS Solr highlight}
\usage{
itis_highlight(...)
}
\arguments{
\item{...}{Args passed to \code{\link[solrium]{solr_highlight}}}
}
\description{
ITIS Solr highlight
}
\examples{
\dontrun{
itis_highlight(q = "rank:Species", hl.fl = 'rank', rows=10)
}
}

