% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itis.R
\name{last_change_date}
\alias{last_change_date}
\title{Provides the date the ITIS database was last updated.}
\usage{
last_change_date(wt = "json", raw = FALSE, ...)
}
\arguments{
\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Provides the date the ITIS database was last updated.
}
\examples{
\dontrun{
last_change_date(config=timeout(3))
}
}

