% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_year.R
\name{review_year}
\alias{review_year}
\title{Returns the review year for the TSN.}
\usage{
review_year(tsn, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{tsn}{TSN for a taxonomic group (numeric). Required.}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: `FALSE`}
}
\value{
a data.frame
}
\description{
Returns the review year for the TSN.
}
\examples{
\dontrun{
review_year(tsn = 180541)
}
}
