% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_anymatch.R
\name{search_anymatch}
\alias{search_anymatch}
\title{Search for any match}
\usage{
search_anymatch(x, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search for any match
}
\examples{
\dontrun{
search_anymatch(x=202385, config=timeout(3))
search_anymatch(x="dolphin", config=timeout(3))
}
}

