% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_scientific.R
\name{search_scientific}
\alias{search_scientific}
\title{Search by scientific name}
\usage{
search_scientific(x, wt = "json", raw = FALSE, ...)
}
\arguments{
\item{x}{text or taxonomic serial number (TSN) (character or numeric)}

\item{wt}{(character) One of "json" or "xml". Required.}

\item{raw}{(logical) Return raw JSON or XML as character string. Required.
Default: \code{FALSE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search by scientific name
}
\examples{
\dontrun{
search_scientific("Tardigrada", config=timeout(3))
search_scientific("Quercus douglasii", config=timeout(3))
}
}

