% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solr.R
\name{solr}
\alias{solr}
\title{ITIS Solr Methods}
\description{
ITIS provides access to their data via their Solr service described at
\url{https://www.itis.gov/solr_documentation.html}.  This is a powerful
interace to ITIS data as you have access to a very flexible query
interface.
}
\section{Functions}{

\itemize{
\item \code{\link[=itis_search]{itis_search()}} - Search
\item \code{\link[=itis_group]{itis_group()}} - Group
\item \code{\link[=itis_highlight]{itis_highlight()}} - Highlight
\item \code{\link[=itis_facet]{itis_facet()}} - Facet
}
}

